+++ src/impls/arrayvec.rs
@@ -1,7 +1,7 @@
 use arrayvec::ArrayVec;
 use fluent_result::bool::Then;
 
-use crate::{ExceedsCapacity, CollectionError, TryExtend, TryExtendSafe, TryFromIterator};
+use crate::{CollectionError, ExceedsCapacity, TryExtend, TryExtendSafe, TryFromIterator};
 
 /// Tries to create an [`ArrayVec`] from an iterator.
 ///
@@ -12,9 +12,9 @@ use crate::{ExceedsCapacity, CollectionError, TryExtend, TryExtendSafe, TryFromI
 /// ```rust
 #[doc = include_doc::function_body!("tests/doc/arrayvec.rs", try_from_iter_arrayvec_example, [])]
 /// ```
-impl<T, I, const N: usize> TryFromIterator<T, I> for ArrayVec<T, N> 
+impl<T, I, const N: usize> TryFromIterator<T, I> for ArrayVec<T, N>
 where
-    I: IntoIterator<Item = T>
+    I: IntoIterator<Item = T>,
 {
     type Error = CollectionError<T, I::IntoIter, ArrayVec<T, N>, ExceedsCapacity>;
 
@@ -24,17 +24,15 @@ where
         match iter.size_hint() {
             (min, _) if min > N => {
                 let err = ExceedsCapacity::new(N, min);
-                return Err(CollectionError::new(iter, ArrayVec::new(), None, err))
-            },
+                return Err(CollectionError::new(iter, ArrayVec::new(), None, err));
+            }
             (_, Some(max)) if max <= N => return Ok(iter.collect()),
             _ => (),
         }
-        
-        iter.try_fold(ArrayVec::new(), |mut array, item| {
-            match array.try_push(item) {
-                Ok(()) => Ok(array),
-                Err(capacity_err) => Err((array, capacity_err.element())),
-            }
+
+        iter.try_fold(ArrayVec::new(), |mut array, item| match array.try_push(item) {
+            Ok(()) => Ok(array),
+            Err(capacity_err) => Err((array, capacity_err.element())),
         })
         .map_err(|(array, rejected_item)| {
             CollectionError::new(iter, array, Some(rejected_item), ExceedsCapacity::new(N, N + 1))
@@ -44,7 +42,10 @@ where
 
 /// Extends an [`ArrayVec`] with an iterator, failing if the iterator produces more items than the [`ArrayVec`]'s
 /// remaining capacity.
-impl<T, const N: usize> TryExtend<T> for ArrayVec<T, N> {
+impl<T, const N: usize, I> TryExtend<T, I> for ArrayVec<T, N>
+where
+    I: IntoIterator<Item = T>,
+{
     type Error = ExceedsCapacity;
 
     /// Appends an iterator to the [`ArrayVec`], failing if the iterator produces more items than the [`ArrayVec`]'s
@@ -58,10 +59,7 @@ impl<T, const N: usize> TryExtend<T> for ArrayVec<T, N> {
     /// ```rust
     #[doc = include_doc::function_body!("tests/doc/arrayvec.rs", try_extend_arrayvec_example, [])]
     /// ```
-    fn try_extend<I>(&mut self, iter: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = T>,
-    {
+    fn try_extend(&mut self, iter: I) -> Result<(), Self::Error> {
         let mut iter = iter.into_iter();
         let size_guess = iter.size_hint().0;
         (size_guess > self.remaining_capacity()).then_err(ExceedsCapacity::new(N, self.len() + size_guess))?;
@@ -71,7 +69,10 @@ impl<T, const N: usize> TryExtend<T> for ArrayVec<T, N> {
 }
 
 /// Extends an [`ArrayVec`] with strong error guarantee.
-impl<T, const N: usize> TryExtendSafe<T> for ArrayVec<T, N> {
+impl<T, const N: usize, I> TryExtendSafe<T, I> for ArrayVec<T, N>
+where
+    I: IntoIterator<Item = T>,
+{
     /// Appends an iterator to the [`ArrayVec`], failing if the iterator produces more items than the [`ArrayVec`]'s
     /// remaining capacity.
     ///
@@ -82,10 +83,7 @@ impl<T, const N: usize> TryExtendSafe<T> for ArrayVec<T, N> {
     /// ```rust
     #[doc = include_doc::function_body!("tests/doc/arrayvec.rs", try_extend_safe_arrayvec_example, [])]
     /// ```
-    fn try_extend_safe<I>(&mut self, iter: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = T>,
-    {
+    fn try_extend_safe(&mut self, iter: I) -> Result<(), Self::Error> {
         let mut iter = iter.into_iter();
 
         let size_guess = iter.size_hint().0;