+++ src/impls/sets/indexset.rs
@@ -1,20 +1,17 @@
 use std::hash::{BuildHasher, Hash};
 
-use fluent_result::into::IntoResult;
+use fluent_result::bool::Expect;
 use indexmap::IndexSet;
-use size_guess::SizeGuess;
 
-use crate::utils::FoldMut;
-use crate::{CollectionCollision, TryExtend, TryExtendSafe, TryFromIterator, ValueCollision};
+use crate::{CollectionCollision, TryExtend, TryExtendSafe, TryFromIterator};
 
-/// Converts an iterator of values into a [`IndexSet`], failing if a value would collide.
-impl<T: Eq + Hash, I> TryFromIterator<T, I> for IndexSet<T> 
+impl<T: Eq + Hash, I> TryFromIterator<T, I> for IndexSet<T>
 where
-    I: IntoIterator<Item = T>
+    I: IntoIterator<Item = T>,
 {
     type Error = CollectionCollision<T, I::IntoIter, IndexSet<T>>;
 
-    /// Converts an iterator of values into a [`IndexSet`], failing if a value would collide.
+    /// Converts `iter` into a [`IndexSet`], failing if a value would collide.
     ///
     /// See [trait level documentation](trait@TryFromIterator) for an example.
     fn try_from_iter(into_iter: I) -> Result<Self, Self::Error>
@@ -27,7 +24,7 @@ where
         iter.try_fold(IndexSet::with_capacity(size_guess), |mut set, value| match set.contains(&value) {
             true => Err((set, value)),
             false => {
-                _ = set.insert(value);
+                set.insert(value).expect_true("should not be occupied");
                 Ok(set)
             }
         })
@@ -35,53 +32,52 @@ where
     }
 }
 
-/// Appends an iterator of values to the [`IndexSet`], failing if a value would collide.
-impl<T: Eq + Hash, S: BuildHasher> TryExtend<T> for IndexSet<T, S> {
-    type Error = ValueCollision<T>;
+impl<T: Eq + Hash, S: BuildHasher, I> TryExtend<T, I> for IndexSet<T, S>
+where
+    I: IntoIterator<Item = T>,
+{
+    type Error = CollectionCollision<T, I::IntoIter, IndexSet<T>>;
 
-    /// Appends an iterator of values to the set, failing if a value would collide.
-    ///
-    /// This implementation provides a basic error guarantee. If the method returns an error, the
-    /// set may be modified. However, it will still be in a valid state, and the specific
-    /// collision that caused the error will not take effect.
+    /// Extends the set with `iter`, failing if a value would collide, with a basic error guarantee.
     ///
     /// See [trait level documentation](trait@TryExtend) for an example.
-    fn try_extend<I>(&mut self, iter: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = T>,
-    {
-        let iter = iter.into_iter();
-        self.reserve(iter.size_guess());
+    fn try_extend(&mut self, iter: I) -> Result<(), Self::Error> {
+        let mut iter = iter.into_iter();
+        self.reserve(iter.size_hint().0);
 
-        iter.into_iter().try_for_each(|value| match self.contains(&value) {
-            true => ValueCollision::new(value).into_err(),
-            false => Ok(_ = self.insert(value)),
+        iter.try_for_each(|value| match self.contains(&value) {
+            true => Err(value),
+            false => {
+                self.insert(value).expect_true("Should not be occupied");
+                Ok(())
+            },
         })
+        .map_err(|value| CollectionCollision::new(iter, IndexSet::new(), value))
     }
 }
 
-/// Appends an iterator of values to the [`IndexSet`] with a strong error guarantee.
-impl<T: Eq + Hash, S: BuildHasher> TryExtendSafe<T> for IndexSet<T, S> {
-    /// Appends an iterator of values pairs to the [`IndexSet`], failing if a value would collide.
-    ///
-    /// This implementation provides a strong error guarantee. If the method returns an error, the
-    /// set is not modified.
+impl<T: Eq + Hash, S: BuildHasher, I> TryExtendSafe<T, I> for IndexSet<T, S>
+where
+    I: IntoIterator<Item = T>,
+{
+    /// Extends the set with `iter`, erroring if a value would collide, with a strong error guarantee.
     ///
     /// See [trait level documentation](trait@TryExtendSafe) for an example.
-    fn try_extend_safe<I>(&mut self, iter: I) -> Result<(), Self::Error>
-    where
-        I: IntoIterator<Item = T>,
-    {
+    fn try_extend_safe(&mut self, iter: I) -> Result<(), Self::Error> {
         let mut iter = iter.into_iter();
-        let size_guess = iter.size_guess();
+        let size_guess = iter.size_hint().0;
 
-        iter.try_fold_mut(IndexSet::with_capacity(size_guess), |set, value| match self.contains(&value) {
-            true => ValueCollision::new(value).into_err(),
+        iter.try_fold(IndexSet::with_capacity(size_guess), |mut set, value| match self.contains(&value) {
+            true => Err((set, value)),
             false => match set.contains(&value) {
-                true => ValueCollision::new(value).into_err(),
-                false => Ok(_ = set.insert(value)),
+                true => Err((set, value)),
+                false => {
+                    set.insert(value).expect_true("should not be occupied");
+                    Ok(set)
+                }
             },
         })
         .map(|set| self.extend(set))
+        .map_err(|(set, value)| CollectionCollision::new(iter, set, value))
     }
 }