+++ src/utils/mod.rs
@@ -3,7 +3,10 @@ mod fixed_size_hint;
 mod fold_mut;
 mod identifiable;
 
-pub(crate) use debug_option::DebugOption;
+pub(crate) use debug_option::OptionTypeDebug;
+#[allow(unused_imports)]
 pub use fixed_size_hint::{FixedSizeHint, FixedSizeHintEx};
+#[allow(unused_imports)]
 pub use fold_mut::FoldMut;
+#[allow(unused_imports)]
 pub use identifiable::Identifiable;